requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot==20.4
requests
